+++ src/WebLinterTest/Settings.cs
@@ -28,5 +28,6 @@ namespace WebLinterTest
         public bool CssLintEnable { get; set; } = true;
         public bool ESLintEnable { get; set; } = true;
         public bool TSLintEnable { get; set; } = true;
+        public bool TSLintWarningsAsErrors { get; set; }
     }
 }